Added IBM Watson result, edited Discussion.
Since the IBM Watson result uses a relatively more complex architecture than the one here, it cast doubt on at least some of the speculation in the most recent version of the discussion section. So that's been removed, at least until we get time to absorb the discussion (and particularly the ablation study results) in the IBM Watson paper.